Final changes InshaaAllah

+++ app/src/main/java/com/example/liderstoreagent/data/models/discountsmodel/Discounts.kt
@@ -2,7 +2,7 @@ package com.example.liderstoreagent.data.models.discountsmodel
 
 data class Discounts(
     val deadline: String?,
-    val discount: Double?,
+    val discount: Int?,
     val id: Int,
     val name: String
 )
